+++ haskpy/typeclasses/_applicative.py
@@ -9,10 +9,10 @@ from haskpy.internal import (
 )
 
 # Use the "hidden" module in order to avoid circular imports
-from ._functor import Functor
+from ._apply import Apply
 
 
-class Applicative(Functor):
+class Applicative(Apply):
     """Applicative functor typeclass.
 
     Minimal complete definition::
@@ -41,27 +41,6 @@ class Applicative(Functor):
     def pure(cls, x):
         """a -> m a"""
 
-    def apply(self, f):
-        """m a -> m (a -> b) -> m b
-
-        Default implementation is based on ``apply_to``.
-
-        """
-        return f.apply_to(self)
-
-    def apply_to(self, x):
-        """f (a -> b) -> f a -> f b
-
-        Default implementation is based on ``apply``.
-
-        """
-        return x.apply(self)
-
-    def sequence(self, x):
-        """f a -> f b -> f b"""
-        from haskpy.utils import identity
-        return self.replace(identity).apply_to(x)
-
     def map(self, f):
         """m a -> (a -> b) -> m b
 
@@ -80,34 +59,6 @@ class Applicative(Functor):
         cls = type(self)
         mf = cls.pure(f)
         return self.apply(mf)
-
-    def __matmul__(self, x):
-        """Application operand ``@`` applies similarly as ``<*>`` in Haskell
-
-        ``f @ x`` translates to ``f.apply_to(x)``, ``x.apply(f)`` and
-        ``apply(f, x)``.
-
-        Why ``@`` operator?
-
-        - It's not typically used as often as some other more common operators
-          so less risk for confusion.
-
-        - The operator is not a commutative as isn't ``apply`` either.
-
-        - If we see matrix as some structure, then matrix multiplication takes
-          both left and right operand inside this structure and gives a result
-          also inside this structure, similarly as ``apply`` does. So it's an
-          operator for two operands having a similar structure.
-
-        - The operator evaluates the contained function(s) at the contained
-          value(s). Thus, ``f`` "at" ``x`` makes perfect sense.
-
-        """
-        return self.apply_to(x)
-
-    def __rshift__(self, x):
-        """Sequence with``>>`` similarly as with ``*>`` and ``>>`` in Haskell"""
-        return self.sequence(x)
 
     #
     # Sampling methods for property tests
@@ -138,7 +89,6 @@ class Applicative(Functor):
 
         """
         pass
-
 
     #
     # Test typeclass laws
@@ -242,58 +192,6 @@ class Applicative(Functor):
     #
     # Test laws based on default implementations
     #
-
-    @class_function
-    @assert_output
-    def assert_applicative_apply(cls, u, v):
-        from .applicative import apply
-        return (
-            v.apply(u),
-            apply(u, v),
-            u.apply_to(v),
-        )
-
-    @class_function
-    @given(st.data())
-    def test_applicative_apply(cls, data):
-        # Draw types
-        a = data.draw(testing.sample_eq_type())
-        b = data.draw(testing.sample_type())
-        fa = data.draw(cls.sample_applicative_type(a))
-        fab = data.draw(cls.sample_applicative_type(testing.sample_function(b)))
-
-        # Draw values
-        v = data.draw(fa)
-        u = data.draw(fab)
-
-        cls.assert_applicative_apply(u, v, data=data)
-        return
-
-    @class_function
-    @assert_output
-    def assert_applicative_sequence(cls, u, v):
-        from .applicative import sequence
-        return (
-            Applicative.sequence(u, v),
-            u.sequence(v),
-            sequence(u, v),
-        )
-
-    @class_function
-    @given(st.data())
-    def test_applicative_sequence(cls, data):
-        # Draw types
-        a = data.draw(testing.sample_type())
-        b = data.draw(testing.sample_type())
-        fa = data.draw(cls.sample_applicative_type(a))
-        fb = data.draw(cls.sample_applicative_type(b))
-
-        # Draw values
-        u = data.draw(fa)
-        v = data.draw(fb)
-
-        cls.assert_applicative_sequence(u, v, data=data)
-        return
 
     @class_function
     @assert_output