+++ haskpy/types/compose.py
@@ -185,6 +185,7 @@ def Compose(X, Y):
         )
 
         sample_functor_type = testing.sample_type_from_value()
+        sample_apply_type = sample_functor_type
         sample_applicative_type = sample_functor_type
         sample_monad_type = sample_functor_type