+++ haskpy/types/either.py
@@ -98,6 +98,7 @@ class Either(Monad, Eq):
     sample_functor_type = testing.sample_type_from_value(
         testing.sample_type(),
     )
+    sample_apply_type = sample_functor_type
     sample_applicative_type = sample_functor_type
     sample_monad_type = sample_functor_type