+++ haskpy/types/linkedlist.py
@@ -466,6 +466,7 @@ class LinkedList(Monad, Monoid, Foldable, Eq):
     )
 
     sample_functor_type = testing.sample_type_from_value()
+    sample_apply_type = sample_functor_type
     sample_applicative_type = sample_functor_type
     sample_monad_type = sample_functor_type